Try to split the tests
